+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 #### Bugfixes
 * do not display content of passwords read from env var or file on docker startup, #372
 * fix display errors on early display of import/export page
+* dependency updates for security fixes (elliptic)
 
 #### Enhancements
 * add new docker env vars to load passwords from file (REDIS_PASSWORD_FILE, SENTINEL_PASSWORD_FILE), #364
@@ -13,6 +14,7 @@
 * switch library to display json objects from "json-tree" to "jquery.json-viewer", refs #375
 * add config value and env var to display valid json data as default as formatted json tree object (VIEW_JSON_DEFAULT), refs #375  
 * add config value and env var to disable display of strings as hexadecimal binary data, refs #376
+* add some basic validation to redis connection params given via command line, refs #377
 
 ## Version 0.6.6
 #### Bugfixes